% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_excel.R
\name{scorex_to_xl}
\alias{scorex_to_xl}
\title{Scorex to Excel}
\usage{
scorex_to_xl(x, row = 1, col = 1, wb = NULL, sheet = NULL,
  open = TRUE)
}
\arguments{
\item{x}{An object of class \code{scorex}}

\item{row}{The row to start writing the tables to.
Default is set to 1.}

\item{col}{The column to start writing the tables to.
Default is set to 1.}

\item{wb}{A workbook object created with \code{\link[openxlsx]{createWorkbook}}.
If no is workbook provided one will be created.}

\item{sheet}{A workbook worksheet name created with \code{\link[openxlsx]{addWorksheet}}.
If no worksheet name is provided, a worksheet will be created.}

\item{open}{Should the workbook updated with the dual score tables be opened?
Default is set to \code{TRUE}.}
}
\description{
Output a Scorex object to Excel
}
