% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorex.R
\name{scorex}
\alias{scorex}
\title{scorex}
\usage{
scorex(formula, data, cut_method = c("bins", "breaks", "percentiles"),
  method_args = NULL, exceptions = NULL, custom_cut_fnc = NULL, ...)
}
\description{
A dual score matrix function that allows for multiple
performance, and cross variables.
}
